+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG for FlatCAM beta
 - in Geometry Editor - simplification method is now threaded and optimized
 - in Geometry Editor - fixed a crash on Geometry Editor exit
 - in Geometry Editor - added a new feature: ability to change the orientation (from CW to CCW and the revers) for Polygon and LinearRing geometry elements.
+- in Geometry Editor - added a context menu in the Geometry Table
 
 26.11.2020